refactor(modules): extract "close" static method (aka closePopup, closeModal, closeToast) from main module
