+++ src/staticMethods/index.js
@@ -1,3 +1,4 @@
+export * from './close'
 export * from './set-reset-defaults'
 export * from './misc'
 export * from './dom'
